+++ myems-api/reports/advancedreportfile.py
@@ -56,7 +56,6 @@ class AdvancedReportFileCollection:
             access_control(req)
         else:
             api_key_control(req)
-        print(req.params)
         reporting_period_start_datetime_local = req.params.get('reportingperiodstartdatetime')
         reporting_period_end_datetime_local = req.params.get('reportingperiodenddatetime')