+++ myems-api/reports/combinedequipmentcomparison.py
@@ -39,7 +39,6 @@ class Reporting:
             access_control(req)
         else:
             api_key_control(req)
-        print(req.params)
         # this procedure accepts combined equipment id or combined equipment uuid to identify a combined equipment
         combined_equipment_id1 = req.params.get("combinedequipmentid1")
         combined_equipment_uuid1 = req.params.get("combinedequipmentuuid1")