+++ myems-api/reports/metersaving.py
@@ -71,7 +71,6 @@ class Reporting:
             access_control(req)
         else:
             api_key_control(req)
-        print(req.params)
         # this procedure accepts meter id or meter uuid to identify a meter
         meter_id = req.params.get('meterid')
         meter_uuid = req.params.get('meteruuid')