+++ myems-api/reports/offlinemeterdaily.py
@@ -61,7 +61,6 @@ class Reporting:
             access_control(req)
         else:
             api_key_control(req)
-        print(req.params)
         offline_meter_id = req.params.get('offlinemeterid')
         offline_meter_uuid = req.params.get('offlinemeteruuid')
         reporting_period_start_datetime_local = req.params.get('reportingperiodstartdatetime')