+++ myems-api/reports/tenantprediction.py
@@ -71,7 +71,6 @@ class Reporting:
             access_control(req)
         else:
             api_key_control(req)
-        print(req.params)
         tenant_id = req.params.get('tenantid')
         tenant_uuid = req.params.get('tenantuuid')
         period_type = req.params.get('periodtype')