+++ myems-api/reports/virtualmetercost.py
@@ -70,7 +70,6 @@ class Reporting:
             access_control(req)
         else:
             api_key_control(req)
-        print(req.params)
         virtual_meter_id = req.params.get('virtualmeterid')
         virtual_meter_uuid = req.params.get('virtualmeteruuid')
         period_type = req.params.get('periodtype')