+++ CHANGELOG.md
@@ -8,6 +8,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 ### Changed
 ### Fixed
 ### Removed
+- removed print(req.params) from myems-api
 
 ## [v6.6.0] - 2026-06-26
 ### Added